+++ media/js/submitFeedbackVue.js
@@ -1,21 +1,20 @@
 console.log('Loaded!')
-;(function() {
+;(function () {
     const vscode = acquireVsCodeApi()
-    const app = new Vue({
-        el: '#app',
-        data: {
+    const app = Vue.createApp({
+        data: () => ({
             comment: '',
             sentiment: '',
             isSubmitting: false,
             error: '',
-        },
+        }),
         mounted() {
-            this.$nextTick(function() {
+            this.$nextTick(function () {
                 window.addEventListener('message', this.handleMessageReceived)
             })
         },
         methods: {
-            handleMessageReceived: function(e) {
+            handleMessageReceived: function (e) {
                 const message = e.data
                 switch (message.statusCode) {
                     case 'Failure':
@@ -25,7 +24,7 @@ console.log('Loaded!')
                         break
                 }
             },
-            submitFeedback: function() {
+            submitFeedback: function () {
                 this.error = ''
                 this.isSubmitting = true
                 console.log('Submitting feedback...')
@@ -37,4 +36,5 @@ console.log('Loaded!')
             },
         },
     })
+    app.mount('#app')
 })()
