+++ package.json
@@ -2131,7 +2131,9 @@
         "@types/xml2js": "^0.4.8",
         "@typescript-eslint/eslint-plugin": "^4.24.0",
         "@typescript-eslint/parser": "^4.29.2",
+        "@vue/compiler-sfc": "^3.2.6",
         "circular-dependency-plugin": "^5.2.2",
+        "css-loader": "^6.2.0",
         "decache": "^4.4.0",
         "esbuild-loader": "2.13.1",
         "eslint": "^7.28.0",
@@ -2159,6 +2161,8 @@
         "vsce": "^1.93.0",
         "vscode-nls-dev": "^3.3.1",
         "vscode-test": "^1.5.2",
+        "vue-loader": "^16.5.0",
+        "vue-style-loader": "^4.1.3",
         "webpack": "^5.38.1",
         "webpack-cli": "^4.7.2"
     },
@@ -2195,7 +2199,7 @@
         "vscode-languageserver": "^6.1.1",
         "vscode-languageserver-textdocument": "^1.0.1",
         "vscode-nls": "^4.1.1",
-        "vue": "^2.6.14",
+        "vue": "^3.2.6",
         "winston": "^3.2.1",
         "winston-transport": "^4.3.0",
         "xml2js": "^0.4.19",
