+++ src/test/lambda/vue/samInvoke.test.ts
@@ -4,7 +4,7 @@
  */
 
 import * as assert from 'assert'
-import { finalizeConfig } from '../../../lambda/vue/samInvoke'
+import { finalizeConfig } from '../../../lambda/vue/samInvokeBackend'
 import { AwsSamDebuggerConfiguration } from '../../../shared/sam/debugger/awsSamDebugConfiguration'
 
 describe('Sam Invoke Vue Backend', () => {
