+++ tsconfig.json
@@ -1,6 +1,4 @@
 {
-    "isolatedModules": true,
-    "esModuleInterop": true,
     "compilerOptions": {
         "baseUrl": ".",
         "paths": {
