+++ types/vue-types.d.ts
@@ -1,0 +1,13 @@
+/*!
+ * Copyright 2021 Amazon.com, Inc. or its affiliates. All Rights Reserved.
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+// temporary ambient declaration
+// this does not really extract types from vue components
+// it just prevents TSC from complaining when trying to import them
+declare module '*.vue' {
+    import { defineComponent } from 'vue'
+    const Component: ReturnType<typeof defineComponent>
+    export default Component
+}
